Move WebSocket state management to a separate directory
