+++ optaweb-tsp-ui/src/store/tsp/actions.ts
@@ -14,12 +14,7 @@
  * limitations under the License.
  */
 
-import { Client, Frame } from 'webstomp-client';
 import { ILatLng, ITSPRouteWithSegments } from './types';
-
-// *************************************************************************************************
-// Action types
-// *************************************************************************************************
 
 export enum ActionType {
   SOLUTION_UPDATES_DATA = 'SOLUTION_UPDATES_DATA',
@@ -27,14 +22,7 @@ export enum ActionType {
   ADD_LOCATION = 'ADD_LOCATION',
   LOAD_DEMO = 'LOAD_DEMO',
   CLEAR_SOLUTION = 'CLEAR_SOLUTION',
-  WS_CONNECT = 'WS_CONNECT',
-  WS_CONNECT_SUCCESS = 'WS_CONNECT_SUCCESS',
-  WS_CONNECT_FAILURE = 'WS_CONNECT_FAILURE',
 }
-
-// *************************************************************************************************
-// Action interfaces
-// *************************************************************************************************
 
 export interface IAddLocationAction {
   readonly type: ActionType.ADD_LOCATION;
@@ -59,40 +47,12 @@ export interface IUpdateTSPSolutionAction {
   solution: ITSPRouteWithSegments;
 }
 
-export interface InitWsConnectionAction {
-  readonly type: ActionType.WS_CONNECT;
-  value: string;
-}
-
-export interface IWsConnectionSuccessAction {
-  readonly type: ActionType.WS_CONNECT_SUCCESS;
-  value: Client;
-}
-
-export interface IWsConnectionFailureAction {
-  readonly type: ActionType.WS_CONNECT_FAILURE;
-  value: Frame | CloseEvent;
-}
-
-/**
- * Union type for all actions.
- */
 export type TspAction =
   | IAddLocationAction
   | IClearSolutionAction
   | IDeleteLocationAction
   | IUpdateTSPSolutionAction;
 
-export type WebSocketAction =
-  | InitWsConnectionAction
-  | IWsConnectionFailureAction
-  | IWsConnectionSuccessAction;
-
-// *************************************************************************************************
-// Action creators
-// *************************************************************************************************
-
-// TODO use ActionCreator<IAddLocationAction> for the function interface
 const addLocation = (location: ILatLng): IAddLocationAction => ({
   type: ActionType.ADD_LOCATION,
   value: location,
@@ -118,32 +78,10 @@ const updateTSPSolution = (
   type: ActionType.SOLUTION_UPDATES_DATA,
 });
 
-const initWsConnection = (socketUrl: string): InitWsConnectionAction => ({
-  type: ActionType.WS_CONNECT,
-  value: socketUrl,
-});
-
-const wsConnectionSuccess = (
-  webstompSocket: Client,
-): IWsConnectionSuccessAction => ({
-  type: ActionType.WS_CONNECT_SUCCESS,
-  value: webstompSocket,
-});
-
-const wsConnectionFailure = (
-  err: Frame | CloseEvent,
-): IWsConnectionFailureAction => ({
-  type: ActionType.WS_CONNECT_FAILURE,
-  value: err,
-});
-
 export default {
   addLocation,
   clearSolution,
   deleteLocation,
-  initWsConnection,
   loadDemo,
   updateTSPSolution,
-  wsConnectionFailure,
-  wsConnectionSuccess,
 };
